requirements.txt
appdirs==1.4.4
attrs==22.2.0
black==18.5b0
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==1.4.1
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
facenet-pytorch==2.5.2
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
kiwisolver==1.2.0
matplotlib==3.2.1
networkx==2.4
ninja==1.11.1
numpy==1.18.3
opencv-python==4.5.5.64
pandas==1.1.5
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2023.3
PyWavelets==1.1.1
requests==2.27.1
scikit-image==0.14.5
scikit-learn==0.24.2
scipy==1.4.1
six==1.14.0
threadpoolctl==3.1.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
typing_extensions==4.1.1
urllib3==1.26.15
zipp==3.6.0
